perf:添加accessKey管理

+++ usermgr/user.go
@@ -1,7 +1,9 @@
 package usermgr
 
 import (
+	"database/sql"
 	"fmt"
+	"strings"
 	"time"
 
 	sqlplus "github.com/cheetah-fun-gs/goplus/dao/sql"
@@ -19,28 +21,28 @@ type User struct {
 	Nickname   string           `json:"nickname,omitempty"`
 	Avatar     string           `json:"avatar,omitempty"`
 	Extra      string           `json:"extra,omitempty"`
-	LastLogin  time.Time        `json:"last_login,omitempty"`
-	Created    time.Time        `json:"created,omitempty"`
+	LastLogin  int64            `json:"last_login,omitempty"`
+	Created    int64            `json:"created,omitempty"`
 	Auths      []*UserAuth      `json:"auths,omitempty"`
 	AccessKeys []*UserAccessKey `json:"access_keys,omitempty"`
 }
 
 // UserAuth 第三方认证
 type UserAuth struct {
-	ID        int       `json:"id,omitempty"`
-	AuthName  string    `json:"auth_name,omitempty"`
-	AuthUID   string    `json:"auth_uid,omitempty"`
-	AuthExtra string    `json:"auth_extra,omitempty"`
-	Created   time.Time `json:"created,omitempty"`
+	ID        int    `json:"id,omitempty"`
+	AuthName  string `json:"auth_name,omitempty"`
+	AuthUID   string `json:"auth_uid,omitempty"`
+	AuthExtra string `json:"auth_extra,omitempty"`
+	Created   int64  `json:"created,omitempty"`
 }
 
 // UserAccessKey 访问密钥
 type UserAccessKey struct {
-	ID        int       `json:"id,omitempty"`
-	AccessKey string    `json:"access_key,omitempty"`
-	ExpireAt  time.Time `json:"expire_at,omitempty"`
-	Comment   string    `json:"comment,omitempty"`
-	Created   time.Time `json:"created,omitempty"`
+	ID        int    `json:"id,omitempty"`
+	AccessKey string `json:"access_key,omitempty"`
+	ExpireAt  int64  `json:"expire_at,omitempty"`
+	Comment   string `json:"comment,omitempty"`
+	Created   int64  `json:"created,omitempty"`
 }
 
 // Login 登录
@@ -138,15 +140,25 @@ func (user *User) BindAuth(authName, authUID, authExtra string) error {
 		AuthName:  authName,
 		AuthUID:   authUID,
 		AuthExtra: authExtra,
-		Created:   now,
+		Created:   now.Unix(),
 	})
 	return nil
 }
 
 // UnbindAuth 解绑第三方认证
 func (user *User) UnbindAuth(authName string) error {
-	query := fmt.Sprintf("DELETE FROM %v WHERE uid = ? AND auth_name = ?;",
-		user.mgr.tableUserAuth.Name)
+	var isMatch bool
+
+	for _, v := range user.Auths {
+		if v.AuthName == authName {
+			isMatch = true
+		}
+	}
+	if !isMatch {
+		return fmt.Errorf("auth not found: %v", authName)
+	}
+
+	query := fmt.Sprintf("DELETE FROM %v WHERE uid = ? AND auth_name = ?;", user.mgr.tableUserAuth.Name)
 	args := []interface{}{user.UID, authName}
 	updateCount, err := sqlplus.RowsAffected(user.mgr.db.Exec(query, args...))
 	if err != nil {
@@ -167,11 +179,30 @@ func (user *User) UnbindAuth(authName string) error {
 }
 
 // UpdateInfo 更新用户信息
-func (user *User) UpdateInfo(nickname, avatar, extra string) error {
+func (user *User) UpdateInfo(nickname, avatar, extra *string) error {
+	if nickname == nil && avatar == nil && extra == nil {
+		return fmt.Errorf("no valid params")
+	}
+
+	splits := []string{}
+	args := []interface{}{}
+	if nickname != nil {
+		splits = append(splits, "nickname = ?")
+		args = append(args, nickname)
+	}
+	if avatar != nil {
+		splits = append(splits, "avatar = ?")
+		args = append(args, avatar)
+	}
+	if extra != nil {
+		splits = append(splits, "extra = ?")
+		args = append(args, extra)
+	}
+
 	now := time.Now()
-	query := fmt.Sprintf("UPDATE %v Set nickname = ?, avatar = ?, extra = ?, updated = ? WHERE id = ?;",
-		user.mgr.tableUser.Name)
-	args := []interface{}{nickname, avatar, extra, now, user.ID}
+	query := fmt.Sprintf("UPDATE %v Set %v, updated = ? WHERE id = ?;",
+		strings.Join(splits, ", "), user.mgr.tableUser.Name)
+	args = append(args, now, user.ID)
 	updateCount, err := sqlplus.RowsAffected(user.mgr.db.Exec(query, args...))
 	if err != nil {
 		return err
@@ -180,9 +211,15 @@ func (user *User) UpdateInfo(nickname, avatar, extra string) error {
 		return ErrorNotFound
 	}
 
-	user.Nickname = nickname
-	user.Avatar = avatar
-	user.Extra = extra
+	if nickname != nil {
+		user.Nickname = *nickname
+	}
+	if avatar != nil {
+		user.Avatar = *avatar
+	}
+	if extra != nil {
+		user.Extra = *extra
+	}
 	return nil
 }
 
@@ -335,5 +372,125 @@ func (user *User) UpdatePasswordWithPassword(oldRawPassword, newRawPassword stri
 	if updateCount == 0 {
 		return ErrorNotFound
 	}
+	return nil
+}
+
+// GenerateAccessKey 生成一个 access key
+func (user *User) GenerateAccessKey(comment string, expireAts ...time.Time) (*UserAccessKey, error) {
+	expireAt := sql.NullTime{}
+	if len(expireAts) > 0 {
+		expireAt.Valid = true
+		expireAt.Time = expireAts[0]
+	}
+	accessKey := user.mgr.generateAccessKey()
+	now := time.Now()
+	data := &ModelUserAccessKey{
+		AccessKey: accessKey,
+		UID:       user.UID,
+		Comment:   comment,
+		ExpireAt:  expireAt,
+		Created:   now,
+		Updated:   now,
+	}
+	query, args := sqlplus.GenInsert(user.mgr.tableUserAccessKey.Name, data)
+	aid, err := sqlplus.LastInsertId(user.mgr.db.Exec(query, args...))
+	if err != nil {
+		return nil, err
+	}
+
+	userAccessKey := &UserAccessKey{
+		ID:        int(aid),
+		AccessKey: accessKey,
+		Comment:   comment,
+		Created:   now.Unix(),
+	}
+	if expireAt.Valid {
+		userAccessKey.ExpireAt = expireAt.Time.Unix()
+	}
+	user.AccessKeys = append(user.AccessKeys, userAccessKey)
+	return userAccessKey, nil
+}
+
+// UpdateAccessKey 更新一个 access key
+func (user *User) UpdateAccessKey(accessKeyID int, comment *string, expireAt *time.Time) error {
+	var userAccessKey *UserAccessKey
+
+	for _, v := range user.AccessKeys {
+		if v.ID == accessKeyID {
+			userAccessKey = v
+		}
+	}
+
+	if userAccessKey == nil {
+		return fmt.Errorf("accessKey not found: %v", accessKeyID)
+	}
+
+	if comment == nil && expireAt == nil {
+		return fmt.Errorf("no valid params")
+	}
+
+	splits := []string{}
+	args := []interface{}{}
+	if comment != nil {
+		splits = append(splits, "comment = ?")
+		args = append(args, comment)
+	}
+	if expireAt != nil {
+		splits = append(splits, "expire_at = ?")
+		args = append(args, expireAt)
+	}
+
+	now := time.Now()
+	query := fmt.Sprintf("UPDATE %v Set %v, updated = ? WHERE id = ?;",
+		strings.Join(splits, ", "), user.mgr.tableUserAccessKey.Name)
+	args = append(args, now, accessKeyID)
+	updateCount, err := sqlplus.RowsAffected(user.mgr.db.Exec(query, args...))
+	if err != nil {
+		return err
+	}
+	if updateCount == 0 {
+		return ErrorNotFound
+	}
+
+	if comment != nil {
+		userAccessKey.Comment = *comment
+	}
+	if expireAt != nil {
+		userAccessKey.ExpireAt = (*expireAt).Unix()
+	}
+	return nil
+}
+
+// DeleteAccessKey 更新一个 access key
+func (user *User) DeleteAccessKey(accessKeyID int) error {
+	var isMatch bool
+
+	for _, v := range user.AccessKeys {
+		if v.ID == accessKeyID {
+			isMatch = true
+		}
+	}
+
+	if !isMatch {
+		return fmt.Errorf("accessKey not found: %v", accessKeyID)
+	}
+
+	query := fmt.Sprintf("DELETE FROM %v WHERE id = ?;", user.mgr.tableUserAccessKey.Name)
+	args := []interface{}{accessKeyID}
+	updateCount, err := sqlplus.RowsAffected(user.mgr.db.Exec(query, args...))
+	if err != nil {
+		return err
+	}
+	if updateCount == 0 {
+		return ErrorNotFound
+	}
+
+	accessKeys := []*UserAccessKey{}
+	for _, v := range user.AccessKeys {
+		if v.ID != accessKeyID {
+			accessKeys = append(accessKeys, v)
+		}
+	}
+	user.AccessKeys = accessKeys
 	return nil
 }
